+++ ton_sdk/src/tests/test_local_tvm.rs
@@ -58,7 +58,7 @@ fn test_local_call_accept_error() {
         "{}".to_owned(),
         PIGGY_BANK_CONTRACT_ABI.to_owned(),
         None,
-        None);
+        Default::default());
     assert!(result.is_err());
 }
 
@@ -73,7 +73,7 @@ fn test_executor_call() {
         "{\"to\": \"0:e6392da8a96f648098f818501f0211f27c89675e5f196445d211947b48e7c85b\"}".to_owned(),
         PIGGY_BANK_CONTRACT_ABI.to_owned(),
         Some(&keypair),
-        None).expect("Error calling contract");
+        Default::default()).expect("Error calling contract");
     assert!(result.transaction.out_messages.len() == 1);
     assert!(!result.transaction.aborted);